Move upgrades for canvasitem parser items to separate module
They're basically legacy, now all items are at top level.
